+++ src/tiledapp/main.cpp
@@ -550,6 +550,12 @@ int main(int argc, char *argv[])
                 return 1;
             }
             Preferences::setStartupProject(filePath);
+        } if (fileInfo.suffix() == QLatin1String("tiled-session")) {
+            if (!fileInfo.exists()) {
+                qWarning().noquote() << QCoreApplication::translate("Command line", "Session file '%1' not found.").arg(fileName);
+                return 1;
+            }
+            Preferences::setStartupSession(filePath);
         } else {
             filesToOpen.append(filePath);
         }